feat: implement remove cart
