+++ src/components/Cart/BasketInfo/BasketInfo.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import AddToCart from '../../DescriptionProduct/AddToCart/AddToCart';
 import styles from './BasketInfo.module.css';
-// import RemoveCart from '../../RemoveCart/RemoveCart';
+import RemoveCart from '../../RemoveCart/RemoveCart';
 import { ICartData } from '../../../interfaces/auth.interface';
 import { formatCentsToCurrency } from '../../../utils/format-to-cents';
 
@@ -60,13 +60,13 @@ export default function BasketInfo({ data }: ICartData) {
                 marginBottom: '5px',
                 fontSize: '0.8rem',
               }}
-              // onClick={handleApplyPromoCode}
+              // onClick={ApplyPromoCode}
             >
               Apply
             </Button>
           ),
         }}
-      />    
+      />
       <Box
         display="flex"
         sx={{
@@ -94,7 +94,7 @@ export default function BasketInfo({ data }: ICartData) {
         <Link to="/catalog" className={styles.link}>
           Continue shopping
         </Link>
-        {/* <RemoveCart /> */}
+        <RemoveCart />
       </Box>
     </>
   );
